requirements.txt
fastapi==0.112.0
mysql_connector_repackaged==0.3.1
nest_asyncio==1.5.6
python-telegram-bot==21.4
uvicorn==0.30.5
